Added new audio files from Asset repo made a new folder called audio in scripts.  moved SoundManager.cs there and added new script to playmusic.
