+++ Assets/Resources/audio/music/apongofballs.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: f1b690be55ec72c49bc2e3c05d291f73
+timeCreated: 1493301429
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 1
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
