+++ Assets/Resources/audio/music/dreamer-dog-funk.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 5e41484277ff0534793c7f0dd9eff6c5
+timeCreated: 1493301452
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 1
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
