+++ Assets/Resources/audio/music/pongballpitchmusic_extended.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: a502b214c0f8440409616df2665fd340
+timeCreated: 1493301416
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 1
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
