+++ Assets/scripts/audio.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d04e1d12c3c026a4fafb6358a4c6b0ff
+folderAsset: yes
+timeCreated: 1493302472
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
