+++ Assets/scripts/audio/MusicPlayer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: fd44898254437fc42b610fbb15767d93
+timeCreated: 1493302636
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
