+++ Assets/scripts/audio/SoundManager.cs
@@ -1,0 +1,88 @@
+﻿/*Original Source code License
+[Copyright 2015 Unity Technologies
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+
+See the License for the specific language governing permissions and
+limitations under the License]
+
+Additional Changes Copyright Stewart Martin with PongBall team
+*/
+
+
+using UnityEngine;
+using System.Collections;
+
+
+public class SoundManager : MonoBehaviour 
+{
+	public AudioSource efxSource;					//Drag a reference to the audio source which will play the sound effects.
+	public AudioSource musicSource;					//Drag a reference to the audio source which will play the music.
+	public static SoundManager instance = null;		//Allows other scripts to call functions from SoundManager.				
+	public float lowPitchRange = .95f;				//The lowest a sound effect will be randomly pitched.
+	public float highPitchRange = 1.05f;			//The highest a sound effect will be randomly pitched.
+    public bool trueForKyle3Wynn;
+	
+	void Awake ()
+	{
+		//Check if there is already an instance of SoundManager
+		if (instance == null)
+			//if not, set it to this.
+			instance = this;
+		//If instance already exists:
+		else if (instance != this)
+			//Destroy this, this enforces our singleton pattern so there can only be one instance of SoundManager.
+			Destroy (gameObject);
+		
+		//Set SoundManager to DontDestroyOnLoad so that it won't be destroyed when reloading our scene.
+		DontDestroyOnLoad (gameObject);
+	}
+
+	//Used to play single sound clips.
+	public void PlaySingle(AudioClip clip)
+	{
+		//Set the clip of our efxSource audio source to the clip passed in as a parameter.
+		efxSource.clip = clip;
+		
+		//Play the clip.
+		efxSource.Play ();
+	}
+	
+	
+	//RandomizeSfx chooses randomly between various audio clips and slightly changes their pitch.
+	public void RandomizeSfx (params AudioClip[] clips)
+	{
+		//Generate a random number between 0 and the length of our array of clips passed in.
+		int randomIndex = Random.Range(0, clips.Length);
+		
+		//Choose a random pitch to play back our clip at between our high and low pitch ranges.
+		float randomPitch = Random.Range(lowPitchRange, highPitchRange);
+		
+		//Set the pitch of the audio source to the randomly chosen pitch.
+		efxSource.pitch = randomPitch;
+		
+		//Set the clip to the clip at our randomly chosen index.
+		efxSource.clip = clips[randomIndex];
+		
+		//Play the clip.
+		efxSource.Play();
+	}
+
+    //like RandomizeSFX minus the random pitch
+    public void RandomSFX(params AudioClip[] clips)
+    {
+        int randomIndex = Random.Range(0, clips.Length);
+        efxSource.clip = clips[randomIndex];
+        efxSource.Play();
+
+    }
+}
+
