+++ Assets/scripts/audio/SoundManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: ba934ac568fc24da98f198aca7c2d107
+timeCreated: 1492506181
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
