+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 5.6.0xf3Linux
+m_EditorVersion: 5.6.0f3
